Deployment/requirements.txt
onnxruntime
numpy
Pillow
fastapi[standard]
uvicorn